feat: uwsgi logvar mapper
This patchset adds middleware to capture information from the WSGI
environment and log it using uwsgi.

Given the specific nature of this module and the size of the uwsgi
dependency, the decision was made to not include uwsgi in the
requirements file. The current solution is to separate all
uwsgi-dependent modules into an eom/uwsgi directory, and have an
import-time guard in the __init__ file that advises users
appropriately.

A simple unit test was added to test the guard functionality.

Due to the nature of the mapper module, a functional test was added to
exercise its functionality. This functional test launches a uwsgi
server process with a trivial app and explores the contents of the
uwsgi server output. As a result, uwsgi was added to
test-requirements.

Since pypy and uwsgi are difficult to make work together, that test
suite has been decorated to skip those tests.