feat: is_slow test helper
This patchset adds a test helper decorator that allows one to annotate
tests as "slow". Slow tests will not be executed by a test runner
unless the EOM_TEST_SLOW environment variable is nonempty.

The *_limit tests for governor.py were decorated, as each of these was
taking >30 seconds to run.