feat(*): python 3 support
Most of eom was already python 3 compatible. The fixes involved the
following files:

* eom/__init__.py - handle gettext correctly in python 3
* tests/functional/test_governor - a lot

The governor tests were incorrectly setting daemon to True for the
HTTP server process. multiprocessing.Process is already a
subprocess. Furthermore, a few changes had to be made to handle
string/bytes and the new division semantics for Python 3.