Enhancement: Operator specified max cache time for EOM Auth
EOM Auth was storing the tokens for the remaining life of a token.
It is useful, however, that the token be stored for only a short time period that is operator defined.

If the operator chooses to add the setting, then EOM Auth will honor the setting.
If not, EOM Auth will behave as before and use the token's expiration time.