Enhancement: Added a check on the base64 decoding
There have been issues with the base64 encode/decode functionality that even with the prior commits in this branch (bugfix_auth_character_encoding) don't seem to be fully resolved.
Further, the py2 information seems to indicate that it is not possible to 100% solve the problem. Therefore, to reduce the occurrence in product we're going to try to decode it immediately after encoding it in
order to try to catch the error and log an appropriate message. This will also cause the call to fail in EOM Middleware instead of in the WSGI Application.